Restore vm_size_t to natural_t
Plenty of places such as glibc RPC uses currently assume that vm_size_t is a
natural_t (thus int on 32bit), not an unsigned long.

* i386/include/mach/i386/vm_types.h (vm_size_t) [!__x86_64__]: Set type
to natural_t.
* device/dev_page.c (device_pager_data_request): Cast vm_size_t to unsigned
long.
* i386/i386at/model_dep.c (c_boot_entry): Likewise.

+++ device/dev_pager.c
@@ -329,7 +329,7 @@ kern_return_t	device_pager_data_request(
 
 	if (device_pager_debug)
 		printf("(device_pager)data_request: pager=%p, offset=0x%lx, length=0x%lx\n",
-			pager, offset, length);
+			pager, (unsigned long) offset, (unsigned long) length);
 
 	ds = dev_pager_hash_lookup(pager);
 	if (ds == DEV_PAGER_NULL)
